fix build errors and refactor dashboard imports
